Fix Data Types in HVAC System Functions
This commit fixes the data types of the fields for the following
messages:
- HvacSystemFunctionListDataSelectorsType
- HvacSystemFunctionOperationModeRelationDataType

Before this commit, HvacSystemFunctionOperationModeRelationDataType, had
a single operationModeId for a system function. According to the spec.,
each system function can have multiple operation modes. Also according
to the resource specification, the OperationModeId is a list and not a
single item.
The selctor For HvacSystemFunctionListDataSelectorsType was wrong.
Instead of a single SystemFunctionId, it was a list.

Because of those issues stated above:
1. Sending request for HvacSystemFunctionOperationModeRelationDataType
failed because the datatype was a single item and not a list.
2. Sending request with a selector for HvacSystemFunctionListDataType
failed.

Now, both sending a request for
HvacSystemFunctionOperationModeRelationDataType
and HvacSystemFunctionListDataType with a selector succees.